requirements.txt
# Django Core
Django==5.0.6
django-environ==0.11.2
gunicorn==22.0.0
psycopg2-binary==2.9.9

# Django REST Framework (for APIs)
djangorestframework==3.15.1

# WebSocket/Async Support
channels==4.0.0
channels-redis==4.2.0

# Task Queue
celery==5.4.0
redis==5.0.1

# AI & Machine Learning
google-generativeai==0.7.1
# The following are for RAG features described in agents.md, but not yet implemented.
# sentence-transformers==3.0.1
# faiss-cpu==1.8.0

# LLM Orchestration
langchain==0.3.27
langchain-community==0.3.27

# Microsoft Authentication
msal==1.28.0

# Utilities
Pillow==11.3.0
requests==2.32.3
